requirements.txt
plotly==5.18.0
pandas==2.2.0
markdown==3.5.1
PyYAML==6.0.1
Jinja2==3.1.2